requirements.txt
fastapi
uvicorn
pydantic
requests
cassandra-driver
pandas